initial thoughts on an extended bytecode
with the aim of being able to write repl_fun entirely in bytecode. This
is a very preliminary commit because I immediately run into a problem
giving semantics to the new LexInput and InstallAndRun instructions
because they depend on knowing how tokens will be encoded in CakeML
(i.e. what tags they will use).